requirements.txt
numpy
opencv-python
Pillow
streamlit
torch
torchvision
tqdm
ultralytics
pandas
# The following versions are critical for TensorFlow on Apple Silicon
tensorflow-macos==2.15.0
tensorflow-metal==1.1.0
keras==2.15.0
